honor HAVE_MANPAGES and fix installation

+++ package/dbus/dbus.mk
@@ -55,18 +55,18 @@ $(STAGING_DIR)/usr/lib/libdbus-1.so: $(DBUS_DIR)/$(DBUS_BINARY)
 	$(MAKE) DESTDIR=$(STAGING_DIR) -C $(DBUS_DIR)/dbus install
 
 $(TARGET_DIR)/$(DBUS_TARGET_BINARY): $(STAGING_DIR)/usr/lib/libdbus-1.so
-	mkdir -p $(TARGET_DIR)/var/run/dbus $(TARGET_DIR)/etc/init.d
-	$(MAKE) DESTDIR=$(TARGET_DIR) -C $(DBUS_DIR)/dbus install
-	rm -rf $(TARGET_DIR)/usr/lib/dbus-1.0
-	rm -f $(TARGET_DIR)/usr/lib/libdbus-1.la \
-		$(TARGET_DIR)/usr/lib/libdbus-1.so
-	-$(STRIPCMD) $(STRIP_STRIP_UNNEEDED) $(TARGET_DIR)/usr/lib/libdbus-1.so.3.2.0
-	$(MAKE) DESTDIR=$(TARGET_DIR) initddir=/etc/init.d -C $(DBUS_DIR)/bus install
+	$(INSTALL) -d $(TARGET_DIR)/var/run/dbus $(TARGET_DIR)/etc/init.d
+	$(MAKE) DESTDIR=$(TARGET_DIR) STRIPPROG='$(STRIP)' \
+		initdir=/etc/init.d -C $(DBUS_DIR)/dbus install-strip
+	rm -rf $(TARGET_DIR)/usr/lib/dbus-1.0 \
+		$(TARGET_DIR)/usr/lib/libdbus-1.la \
+		$(TARGET_DIR)/usr/include/dbus-1.0 \
+		$(TARGET_DIR)/usr/lib/pkgconfig
 	$(INSTALL) -m 0755 package/dbus/S97messagebus $(TARGET_DIR)/etc/init.d
 	rm -f $(TARGET_DIR)/etc/init.d/messagebus
+ifneq ($(BR2_HAVE_MANPAGES),y)
 	rm -rf $(TARGET_DIR)/usr/share/man
-	rmdir --ignore-fail-on-non-empty $(TARGET_DIR)/usr/share
-	rm -rf $(TARGET_DIR)/etc/rc.d
+endif
 
 dbus: uclibc expat libxml2-headers $(TARGET_DIR)/$(DBUS_TARGET_BINARY)
